Docs (readme): Add common error messages for *nix
Closes #8

It is common for new or inexperienced developers to face some very common errors when using the module or the compilers (`amxxpc` or `spcomp`). However the error messages are generally obscure and developers might not be able to easily find a solution.

Hence, documentation about these errors and their resolution is added in readme.

Co-authored-by: Joel Timothy Oh <joeltimothyoh@live.com>
Co-authored-by: Leo <leonardjonathanoh@live.com>